+++ lib/src/features/home/presentation/widgets/ticket_card_widget.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:namma_wallet/src/common/helper/date_time_converter.dart';
-import 'package:namma_wallet/src/common/styles/styles.dart';
+import 'package:namma_wallet/src/common/theme/styles.dart';
 import 'package:namma_wallet/src/features/home/domain/generic_details_model.dart';
 
 class TicketCardWidget extends StatelessWidget {
